extract structs to files
