+++ gb-apu/src/channels/sweep.rs
@@ -1,0 +1,10 @@
+#[derive(Default)]
+pub struct Sweep {
+    bits: u8,
+}
+
+impl Sweep {
+    fn sweep_period(&self) {}
+    fn negate(&self) {}
+    fn shift(&self) {}
+}
