+++ gb-apu/src/lib.rs
@@ -1,1 +1,10 @@
 pub mod apu;
+pub mod channels;
+pub mod control;
+
+#[derive(PartialEq, Clone, Copy)]
+pub enum ChannelType {
+    SquareWave,
+    WaveForm,
+    Noise,
+}
